requirements.txt
asgiref==3.7.2
Django==4.2.3
django-nested-admin==4.0.2
djangorestframework==3.14.0
Pillow==10.0.0
python-monkey-business==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
